worker/requirements.txt
kafka-python
aiogram